Larger input buffer and block on connection
The input buffer is now 128 characters and there is no attempt to
handle longer messages. An input flush() function would be useful.

Writes now block until a connection is established. Not only does
this help with the reliabilty of the data connection, but it prevents
data writes from interfering with command (AT) messages.